requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2.1
django-countries==7.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
idna==3.4
jwt==1.3.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.5.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.0
tzdata==2023.3
urllib3==2.0.2
